feat(upstream-by-header) create plugin
This is a custom plugin that will send a request to a different upstream
when the request headers match the user's configured headers. For
example, requests that match route /local will be proxied to the
upstream europe_cluster, except requests that contain the header
X-Country=Italy, which should be proxied to the upstream italy_cluster.

Multiple headers and the upstream name can be configured. For example,
`X-Country=Italy`, `X-Regione=Abruzzo` will go to the upstream
`italy_cluster`, but just just `X-Country=Italy` without the header
`X-Regione` still goes to europe_cluster.

This plugin follows the following plugin template by Kong:
https://github.com/Kong/kong-plugin.