feat(upstream-by-header) header matching algorithm
This commit checks if the request headers match the user's configured
headers by implementing a simple header matching algorithm. If the
matching occurs, the rule's field `upstream_name` will be used as the
upstream hostname for this request.

Related use case: Requests that match the route `/local` will be proxied
to the upstream europe_cluster, except requests that contain the header
X-Country=Italy which should be proxied to upstream italy_cluster.