fix(upstream-by-header) fix matching algorithm
This fixes the problem where multiple request headers are not being
matched to the correct configured rule. In particular, for two
configured headers h1 and h2 as the rule, requests that did not have the
header h1 but had h2 were still being routed to the upstream of that
configured rule.

Improvements to the algorithm could be made to optimize the time
complexity (currently O(r * h), where r is the number of rules and h is
the number of headers for each rule).