tests(upstream-by-header) service upstream routing
This unit test checks if the Service is routing to the configured
upstream plugin by verifying if the response headers have the
`X-Upstream-Name` header name set. If it does, its value must be equal
to the configured `upstream_name` made by the user.

The `X-Upstream-Name` header had to be set manually during the access
phase (on handler.lua).