tests(upstream-by-header) add integration tests
This commit adds more integration tests for the access phase of the
plugin. In total, twelve tests were added and they all passed. In
summary: an upstream name and multiple header names and values can be a
configurable rule. Each request with headers that match all the
headers in a rule will be proxied to the `upstream_name` configured for
the rule.

The integration test template was based on  the `kong-plugin` template
repository for new Kong plugins:
https://github.com/Kong/kong-plugin/blob/master/spec/myplugin/01-access_spec.lua