chore(ci) use OPENSSL=1.1.1a to fix Travis build
This commit tries to fix the Travis CI build failing process by setting
the `OPENSSL` environment variable to `1.1.1a`. We are attempting to fix
the build by using the 0.15.0 tag version of the `setup_env.sh` file of
the Kong repository [0]. See this [1] job log for details of the job
error.

[0] https://github.com/Kong/kong/blob/0.15.0/.travis.yml [1]
https://travis-ci.com/murillopaula/kong-upstream-by-header/jobs/194473753