chore(ci) do not run Kong tests for the plugin CI
This commit removes the Kong unit and integration tests as part of the
Travis CI build process for the `upstream-by-header` plugin. Since the
plugin itself does not modify any of the Kong core functionalities (it
only extends it!), this should not be an issue.

See this Travis CI job for details:
https://travis-ci.com/murillopaula/kong-upstream-by-header/jobs/194475225