docs(upstream-by-header) update todo section
In this commit, we are updating the TODO section to reflect the changes made to the Travis CI configuration file [0] that are now making the build pass. We also point out the future improvements that could be made to the current Travis CI build settings. Also in the TODO section, we rewrote the explanation about the current header matching algorithm and future improvements, by making a better use of algorithm analysis terminology.

[0] https://github.com/murillopaula/kong-upstream-by-header/commit/9d839a9276129903d482e710daf3eaadb0c63bd4